Update: 优化下载过期时间，优化资源文件清单校验信息

+++ Assets/CosmosFramework/Runtime/Modules/Download/IDownloadManager.cs
@@ -50,7 +50,7 @@ namespace Cosmos.Download
         /// <summary>
         /// 任务过期时间，以秒为单位；
         /// </summary>
-        float DownloadTimeout { get; set; }
+        int DownloadTimeout { get; set; }
         /// <summary>
         /// 是否正在下载；
         /// </summary>
